Primeiro Commit

+++ shiva.sql
@@ -8,9 +8,8 @@ create table Eventos(
     dT date not null,
     hora varchar(45) not null
 )Engine=InnoDB;
-
 create table Cliente(
-	CPF int auto_increment primary key not null,
+	CPF bigint auto_increment primary key not null,
     Nome varchar(100) not null,
     email varchar(100) not null,
     telefone varchar(14) not null,
@@ -18,13 +17,12 @@ create table Cliente(
 )Engine=InnoDB;
 
 select * from Cliente;
-
 -- relacionameto
 CREATE TABLE EventoCliente(
   idEventoCliente INT NOT NULL AUTO_INCREMENT primary key
 )ENGINE = InnoDB;
 
-alter table EventoCliente add cliente_codigo int not null;
+alter table EventoCliente add cliente_codigo bigint not null;
 alter table EventoCliente add codigo_evento int not null;
 
 alter table EventoCliente add constraint evCliente
